# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateAccept": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "conditions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 12,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "exceptions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 12,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exceptions"
          },
          "expected_version": {
            "exclusiveMinimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "suggested_destination_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Destination Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "body",
          "expected_version"
        ],
        "title": "CandidateAccept",
        "type": "object"
      },
      "CandidateAtomicity": {
        "enum": [
          "atomic",
          "needs_split"
        ],
        "title": "CandidateAtomicity",
        "type": "string"
      },
      "CandidateEdit": {
        "additionalProperties": false,
        "properties": {
          "atomicity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateAtomicity"
              },
              {
                "type": "null"
              }
            ]
          },
          "body": {
            "anyOf": [
              {
                "maxLength": 20000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "conditions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 12,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "evidence_section_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 12,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Section Ids"
          },
          "exceptions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 12,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exceptions"
          },
          "expected_version": {
            "exclusiveMinimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "suggested_destination_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Destination Id"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "CandidateEdit",
        "type": "object"
      },
...
      "CandidateNeedsVerification": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "conditions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 12,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "exceptions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 12,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exceptions"
          },
          "expected_version": {
            "exclusiveMinimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "suggested_destination_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Destination Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "body",
          "expected_version",
          "reason"
        ],
        "title": "CandidateNeedsVerification",
        "type": "object"
      },
      "CandidateReject": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "conditions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 12,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "exceptions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 12,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exceptions"
          },
          "expected_version": {
            "exclusiveMinimum": 0.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "suggested_destination_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Destination Id"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "body",
          "expected_version"
        ],
        "title": "CandidateReject",
        "type": "object"
      },
...
      "CandidateReviewRequestResponse": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/CandidateReviewResponse"
          },
          "status": {
            "default": "succeeded",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key",
          "action",
          "result"
        ],
        "title": "CandidateReviewRequestResponse",
        "type": "object"
      },
      "CandidateReviewResponse": {
        "properties": {
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "candidate_review_id": {
            "format": "uuid",
            "title": "Candidate Review Id",
            "type": "string"
          },
          "candidate_status": {
            "title": "Candidate Status",
            "type": "string"
          },
          "candidate_version": {
            "title": "Candidate Version",
            "type": "integer"
          },
          "evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "knowledge_node_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Node Id"
          },
          "knowledge_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Revision Id"
          },
          "result_id": {
            "format": "uuid",
            "title": "Result Id",
            "type": "string"
          },
          "review_card_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Card Id"
          }
        },
        "required": [
          "result_id",
          "candidate_review_id",
          "candidate_id",
          "candidate_version",
          "candidate_status",
          "knowledge_node_id",
          "knowledge_revision_id",
          "review_card_id",
          "evidence_ids"
        ],
        "title": "CandidateReviewResponse",
        "type": "object"
      },
...
      "DestinationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DestinationResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DestinationListResponse",
        "type": "object"
      },
      "DestinationResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "path": {
            "items": {
              "type": "string"
            },
            "title": "Path",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "DestinationResponse",
        "type": "object"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "D5 immutable source ingestion, parsing, structured AI extraction, and human candidate review API for the personal knowledge workbench.",
    "title": "Zixu Knowledge Workbench API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge-spaces/{space_id}/candidates/{candidate_id}": {
      "patch": {
        "operationId": "edit_extraction_candidate",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Edit Candidate",
        "tags": [
          "candidate-review"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/candidates/{candidate_id}/accept": {
      "post": {
        "operationId": "accept_extraction_candidate",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateAccept"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Accept Candidate",
        "tags": [
          "candidate-review"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/candidates/{candidate_id}/mark-needs-verification": {
      "post": {
        "operationId": "mark_candidate_needs_verification",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateNeedsVerification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Mark Candidate Needs Verification",
        "tags": [
          "candidate-review"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/candidates/{candidate_id}/reject": {
      "post": {
        "operationId": "reject_extraction_candidate",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateReject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Reject Candidate",
        "tags": [
          "candidate-review"
        ]
      }
    },
    "/api/v1/knowledge-spaces/{space_id}/candidates/{candidate_id}/review-requests/{idempotency_key}": {
      "get": {
        "operationId": "get_candidate_review_request",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "idempotency_key",
            "required": true,
            "schema": {
              "title": "Idempotency Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateReviewRequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Candidate Review Request",
        "tags": [
          "candidate-review"
        ]
      }
    },
...
    "/api/v1/knowledge-spaces/{space_id}/knowledge-destinations": {
      "get": {
        "operationId": "list_knowledge_destinations",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Destinations",
        "tags": [
          "candidate-review"
        ]
      }
    },
...
  }
}